+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
+## [Unreleased]
+### Changed
+- Shinobi controller removes all temp directories.
+
 ## 3.0.2
 ### Changed
 - Aligned detection of modified monitor configuration closer to Shinobi's behaviour.